Adds inspected flag and related endpoints for candidate pages
Introduces an 'inspected' boolean property to candidate pages to track inspection status, including DTO and database changes. Adds endpoints to update the inspection status and to move candidate pages to webpages while cleaning up. Enhances logic to only process uninspected, matched candidates for specific operations.

Improves traceability and state management for candidate pages.

+++ src/webpage/webpage.service.ts
@@ -87,6 +87,7 @@ export class WebpageService {
     }
 
     const entity = this.webpagesRepository.create(createWebpageDto);
+
     const shopProductEntity = await this.shopProductService.findOneByProductId(
       createWebpageDto.productId,
       createWebpageDto.shopId,
@@ -134,9 +135,12 @@ export class WebpageService {
     });
     console.log(`Page being created: ${createWebpageDto.url}`);
     console.log(webpageEntity);
-    await this.shopProductService.removeCandidatePageFromShopProduct(
-      webpageEntity.url,
-    );
+    if (shopProductEntity.candidatePages) {
+      await this.shopProductService.removeCandidatePageFromShopProduct(
+        webpageEntity.url,
+      );
+    }
+
     return webpageEntity;
   }
